requirements.txt
numpy
scipy
h5py
pymp-pypi
pyscf
